COL-1784, whiteboard: wire up the 'move' and 'draw' tools
